+++ packages/alfa-rules/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @siteimprove/alfa-rules
 
+## 0.111.0
+
 ## 0.110.0
 
 ### Minor Changes